# docs/swagger.yaml
...
openapi: 3.1.0
paths:
  /admin/centre/submissions:
    get:
      parameters:
      - name: status
        in: query
        schema:
          type: string
          default: SUBMITTED
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Centre Submissions
      tags:
      - Admin Resource
      security:
      - BasicAuth:
        - COACH
        - ADMIN
  /admin/centre/users:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Centre Users
      tags:
      - Admin Resource
      security:
      - BasicAuth:
        - COACH
        - ADMIN
  /admin/submissions/{submissionId}/approve:
    post:
      parameters:
      - name: submissionId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitChallengeResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Approve
      tags:
      - Admin Resource
      security:
      - BasicAuth:
        - COACH
        - ADMIN
  /admin/submissions/{submissionId}/reject:
    post:
      parameters:
      - name: submissionId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitChallengeResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Reject
      tags:
      - Admin Resource
      security:
      - BasicAuth:
        - COACH
        - ADMIN
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Bad Request
      summary: Login
      tags:
      - Auth Resource
  /auth/logout:
    post:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Logout
      tags:
      - Auth Resource
      security:
      - BasicAuth:
        - PLAYER
        - PARENT
        - COACH
        - ADMIN
  /auth/me:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Me
      tags:
      - Auth Resource
      security:
      - BasicAuth:
        - PLAYER
        - PARENT
        - COACH
        - ADMIN
...
  /leaderboards/centre:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Centre Leaderboard
      tags:
      - Leaderboard Resource
      security:
      - BasicAuth:
        - PLAYER
        - PARENT
        - COACH
        - ADMIN
  /leaderboards/global:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Global Leaderboard
      tags:
      - Leaderboard Resource
      security:
      - BasicAuth:
        - PLAYER
        - PARENT
        - COACH
        - ADMIN
...
  /me/badges:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadgesResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Badges
      tags:
      - Player Resource
      security:
      - BasicAuth:
        - PLAYER
        - PARENT
        - COACH
        - ADMIN
...
components:
  schemas:
...
    BadgeDto:
      type: object
      properties:
        badgeCategory:
          type: string
        xp:
          type: integer
          format: int32
        level:
          type: string
        nextLevelAtXp:
          type: integer
          format: int32
    BadgesResponse:
      type: object
      properties:
        badges:
          type: array
          items:
            $ref: "#/components/schemas/BadgeDto"
...
    LeaderboardResponse:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardRow"
    LeaderboardRow:
      type: object
      properties:
        rank:
          type: integer
          format: int32
        username:
          type: string
        displayName:
          type: string
        totalXp:
          type: integer
          format: int32
        centreName:
          type: string
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UUID"
        username:
          type: string
        role:
          type: string
        centreId:
          $ref: "#/components/schemas/UUID"
        displayName:
          type: string
...
    ReviewRequest:
      type: object
      properties:
        reviewerComment:
          type: string
    SubmissionDto:
      type: object
      properties:
        submissionId:
          $ref: "#/components/schemas/UUID"
        challengeId:
          $ref: "#/components/schemas/UUID"
        challengeName:
          type: string
        username:
          type: string
        displayName:
          type: string
        noteText:
          type: string
        submittedAt:
          type: string
    SubmissionsResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionDto"
...
    SubmitChallengeResponse:
      type: object
      properties:
        submissionId:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
    UUID:
      type: string
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
...
    UserDto:
      type: object
      properties:
        userId:
          $ref: "#/components/schemas/UUID"
        centreId:
          $ref: "#/components/schemas/UUID"
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
        active:
          type: boolean
    UsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserDto"
  securitySchemes:
    BasicAuth:
      description: Authentication
      type: http
      scheme: basic
info:
  title: Wishaw YMCA Esports API
  version: 1.0.0
  description: Backend API for the Wishaw YMCA Esports Login App
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value